refactor: restructure letter, new, edit for protected routes
