+++ src/app/(auth-protected)/letter/_components/type-bar/gallery-button.tsx
