+++ src/app/(auth-protected)/letter/_components/type-bar/text-limit.tsx
